+++ publish/CHANGELOG.md
@@ -1,6 +1,7 @@
 - v1.53
   - Fixes state `leave` not working.
   - Fixes parameters `<large>` and `<small>` not working properly. Thanks Kurios.ZeuS!
+  - Fixes poke limit not using offset, instead it was always based on object position.
   - Improves parameters `<calci>`, `<calcf>` and `<calcl>` to support more complex calculations.
   - Removes basic arithmetic support from commands. Use parameters instead.